+++ src/server/server.test.ts
@@ -1,9 +1,14 @@
-var assert = require('assert');
+import app from './server';
+import { equal, ok} from 'assert';
 
 describe('Array', function() {
   describe('#indexOf()', function() {
     it('should return -1 when the value is not present', function() {
-      assert.equal([1, 2, 3].indexOf(4), -1);
+      equal([1, 2, 3].indexOf(4), -1);
     });
+  });
+
+  it('should be defined', () => {
+    ok(app);
   });
 });